pyproject.toml
[project]
name = "insta-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiograpi>=0.0.4",
    "aiopath>=0.7.7",
    "aiosqlite>=0.22.1",
    "beautifulsoup4>=4.14.3",
    "fastapi[standard]>=0.128.7",
    "pydantic-settings>=2.12.0",
    "pyotp>=2.9.0",
    "sqlalchemy>=2.0.46",
    "xai-sdk>=1.6.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
lint = [
    "ruff>=0.15.7",
]

[tool.pytest.ini_options]
pythonpath = "src"

[tool.ruff.lint]
extend-select = ["ASYNC", "FAST", "LOG", "PTH"]